refactor: 야구 게임과 관련 있는 메서드는 BaseBallGame 클래스에 묶어주기
